base/db-requirements.txt
psycopg2-binary==2.8.4
pandas==1.0.3
sqlalchemy==1.3.15
dateparser
dsnparse
